# intermediate/client/openapi.json
{"openapi": "3.0.2", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/items/": {"get": {"tags": ["items"], "summary": "Get Items", "operationId": "get_items", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Item"}}}}}}, "post": {"tags": ["items"], "summary": "Create Item", "operationId": "create_item", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Item"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResponseMessage"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/": {"post": {"tags": ["users"], "summary": "Create User", "operationId": "create_user", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResponseMessage"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "Item": {"title": "Item", "required": ["name", "price"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "price": {"title": "Price", "type": "number"}}}, "ResponseMessage": {"title": "ResponseMessage", "required": ["message"], "type": "object", "properties": {"message": {"title": "Message", "type": "string"}}}, "User": {"title": "User", "required": ["username", "email"], "type": "object", "properties": {"username": {"title": "Username", "type": "string"}, "email": {"title": "Email", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}